+++ src/libmain/shared.hh
@@ -26,7 +26,7 @@ namespace nix {
 Path makeRootName(const Path & gcRoot, int & counter);
 void printGCWarning();
 
-unsigned int getIntArg(const string & opt,
+unsigned long long getIntArg(const string & opt,
     Strings::iterator & i, const Strings::iterator & end);
 
 /* Whether we're running setuid. */
